[Backends] fix potential integer overflow in alloc size (#934)

+++ src/shambackends/include/shambackends/details/memoryHandle.hpp
@@ -64,7 +64,7 @@ namespace sham::details {
      */
     template<USMKindTarget target>
     USMPtrHolder<target> create_usm_ptr(
-        u32 size,
+        size_t size,
         std::shared_ptr<DeviceScheduler> dev_sched,
         std::optional<size_t> alignment = std::nullopt);
